Support previewing of plots from Gonum and Glot
